+++ MOP/src/GameObjects/Others/MopFsmManager.cs
@@ -1,5 +1,4 @@
-﻿using HutongGames.PlayMaker;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace MOP
 {