+++ MOP/src/GameObjects/Vehicles/Vehicle.cs
@@ -1,6 +1,7 @@
 ﻿using MSCLoader;
 using System.Linq;
 using UnityEngine;
+using System.Collections.Generic;
 
 namespace MOP
 {
@@ -23,28 +24,28 @@ namespace MOP
         public Quaternion Rotation { get; set; }
 
         // All objects that cannot be unloaded (because it causes problems) land under that object
-        internal GameObject TemporaryParent;
+        Transform TemporaryParent;
 
-        // List of unloadable objects
-        internal Transform[] AudioObjects;
-        internal Transform FuelTank;    
+        // List of non unloadable objects
+        List<UnloadableObject> unloadableObjects;
 
         // Overwrites the "Component.transform", to prevent eventual mod crashes caused by missuse of Vehicle.transform.
         // Technically, you should use Vehicle.Object.transform (ex. GIFU.Object.Transform), this here just lets you use Vehicle.transform
         // (ex. GIFU.transform).
         public Transform transform => gameObject.transform;
 
-        public Gifu gifuScript;
+        // Loaded only for Satsuma
         public Satsuma satsumaScript;
-        public Fury furyScript;
 
+        // Unity car systems and rigidbody
         internal CarDynamics carDynamics;
         internal Axles axles;
         internal Rigidbody rb;
 
-        bool isHayosiko;
+        // Applies extra fixes, if is set to true.
+        readonly bool isHayosiko;
 
-        bool flatbedScriptActivated;
+        bool flatbedUnloadPreventionActivated;
 
         // Prevents MOP from disabling car's physics when the car has rope hooked
         internal bool IsRopeHooked { get; set; }
@@ -55,41 +56,45 @@ namespace MOP
         /// <param name="gameObjectName"></param>
         public Vehicle(string gameObjectName)
         {
-            // Find the object by name
+            // gameObject the object by name
             gameObject = GameObject.Find(gameObjectName);
 
-            // Dump the object position and rotation
+            // Get the object position and rotation
             Position = gameObject.transform.localPosition;
             Rotation = gameObject.transform.localRotation;
 
             // Creates a new gameobject that is names after the original file + '_TEMP' (ex. "SATSUMA(557kg, 248)_TEMP")
-            TemporaryParent = new GameObject(gameObject.name + "_TEMP");
+            TemporaryParent = new GameObject(gameObject.name + "_TEMP").transform;
+
+            unloadableObjects = new List<UnloadableObject>();
 
             // Get the object's child which are responsible for audio
-            AudioObjects = GetAudioObjects();
+            foreach (Transform audioObject in FindAudioObjects())
+            {
+                unloadableObjects.Add(new UnloadableObject(audioObject));
+            }
 
             // Fix for fuel level resetting after respawn
-            FuelTank = FindFuelTank();
+            Transform fuelTank = gameObject.transform.Find("FuelTank");
+            if (fuelTank != null)
+            {
+                unloadableObjects.Add(new UnloadableObject(fuelTank));
+            }
+
+            // If the vehicle is Gifu, find knobs and add them to list of unloadableObjects
+            if (gameObject.name == "GIFU(750/450psi)")
+            {
+                unloadableObjects.Add(new UnloadableObject(gameObject.transform.Find("Dashboard").Find("Knobs")));
+            }
 
             carDynamics = gameObject.GetComponent<CarDynamics>();
             axles = gameObject.GetComponent<Axles>();
             rb = gameObject.GetComponent<Rigidbody>();
 
             if (!gameObject.name.ContainsAny("HAYOSIKO", "FURY"))
-                gameObject.AddComponent<OcclusionObject>();
-
-            Toggle = ToggleActive;
-
-            isHayosiko = gameObject.name.Contains("HAYOSIKO");
-
-            if (isHayosiko && CompatibilityManager.instance.OffroadHayosiko)
             {
-                Toggle = ToggleUnityCar;
+                gameObject.AddComponent<OcclusionObject>();
             }
-
-            // Vehicle is flatbed
-            if (gameObject.name == "FLATBED")
-                FsmHook.FsmInject(transform.Find("Bed/LogTrigger").gameObject, "Add scale", FlatbedSwitchToggleMethod);
 
             // Hook HookFront and HookRear
             // Get hooks first
@@ -110,18 +115,37 @@ namespace MOP
                 FsmHook.FsmInject(hookRear.gameObject, "Activate cable 2", RopeHookUp);
                 FsmHook.FsmInject(hookRear.gameObject, "Remove rope", RopeUnhook);
             }
+
+            // If vehicle is flatbed, hook SwitchToggleMethod to Add scale script
+            if (gameObject.name == "FLATBED")
+            {
+                FsmHook.FsmInject(transform.Find("Bed/LogTrigger").gameObject, "Add scale", FlatbedSwitchToggleMethod);
+            }
+
+            // Set default toggling method - that is entire vehicle
+            Toggle = ToggleActive;
+
+            isHayosiko = gameObject.name.Contains("HAYOSIKO");
+
+            // Fix for Offroad Hayosiko and HayosikoColorfulGauges
+            if (isHayosiko && (CompatibilityManager.instance.OffroadHayosiko || CompatibilityManager.instance.HayosikoColorfulGauges))
+            {
+                Toggle = ToggleUnityCar;
+            }
+
+            // If the vehicle is Fury
+            if (gameObject.name == "FURY(1630kg)" || gameObject.name == "POLICEFERNDALE(1630kg)")
+            {
+                Toggle = ToggleUnityCar;
+            }
+
+            // If the user selected to toggle vehicle's physics only, it overrided any previous set for Toggle method
+            if (MopSettings.ToggleVehiclePhysicsOnly)
+            {
+                Toggle = ToggleUnityCar;
+            }
         }
 
-        /// <summary>
-        /// Retrieves the child audio objects from parent object.
-        /// Basically looks for files with "audio" and "SoundSrc" name in it
-        /// </summary>
-        /// <returns></returns>
-        internal Transform[] GetAudioObjects()
-        {
-            Transform[] childs = gameObject.transform.GetComponentsInChildren<Transform>();
-            return childs.Where(obj => obj.gameObject.name.Contains("audio") || obj.gameObject.name.Contains("SoundSrc")).ToArray();
-        }
 
         public delegate void ToggleHandler(bool enabled);
         public ToggleHandler Toggle;
@@ -147,11 +171,8 @@ namespace MOP
             // We're doing that BEFORE we disable the object.
             if (!enabled)
             {
-                SetParentForChilds(AudioObjects, TemporaryParent);
-                if (FuelTank != null)
-                {
-                    SetParentForChild(FuelTank, TemporaryParent);
-                }
+                for (int i = 0; i < unloadableObjects.Count; i++)
+                    SetParentForChild(unloadableObjects[i].ObjectTransform, TemporaryParent);
 
                 Position = gameObject.transform.localPosition;
                 Rotation = gameObject.transform.localRotation;
@@ -166,11 +187,8 @@ namespace MOP
                 gameObject.transform.localPosition = Position;
                 gameObject.transform.localRotation = Rotation;
 
-                SetParentForChilds(AudioObjects, gameObject);
-                if (FuelTank != null)
-                {
-                    SetParentForChild(FuelTank, gameObject);
-                }
+                for (int i = 0; i < unloadableObjects.Count; i++)
+                    SetParentForChild(unloadableObjects[i].ObjectTransform, unloadableObjects[i].Parent);
             }
         }
 
@@ -180,7 +198,8 @@ namespace MOP
         /// <param name="enabled"></param>
         public void ToggleUnityCar(bool enabled)
         {
-            if (gameObject == null || carDynamics.enabled == enabled || (satsumaScript != null && satsumaScript.IsSatsumaInInspectionArea && enabled == false)) return;
+            if ((gameObject == null) || (carDynamics.enabled == enabled) || (satsumaScript != null && satsumaScript.IsSatsumaInInspectionArea && enabled == false)) 
+                return;
 
             // Prevent disabling car physics if the rope is hooked
             if (IsRopeHooked && gameObject.activeSelf == true) return;
@@ -191,60 +210,44 @@ namespace MOP
         }
 
         /// <summary>
-        /// Changes audio childs parent
-        /// </summary>
-        /// <param name="childs"></param>
-        /// <param name="newParent"></param>
-        internal void SetParentForChilds(Transform[] childs, GameObject newParent)
-        {
-            for (int i = 0; i < childs.Length; i++)
-            {
-                childs[i].parent = newParent.transform;
-            }
-        }
-
-        /// <summary>
         /// Changes parent for single child
         /// </summary>
         /// <param name="child"></param>
         /// <param name="newParent"></param>
-        internal void SetParentForChild(Transform child, GameObject newParent)
+        internal void SetParentForChild(Transform child, Transform newParent)
         {
-            child.transform.parent = newParent.transform;
+            child.transform.parent = newParent;
         }
 
         /// <summary>
-        /// Looks for FuelTank object.
-        /// Returns null if it hasn't been found.
+        /// Retrieves the child audio objects from parent object.
+        /// Basically looks for files with "audio" and "SoundSrc" name in it
         /// </summary>
-        internal Transform FindFuelTank()
+        /// <returns></returns>
+        Transform[] FindAudioObjects()
         {
-            if (gameObject.transform.Find("FuelTank") != null)
-            {
-                return gameObject.transform.Find("FuelTank");
-            }
-
-            return null;
+            Transform[] childs = gameObject.transform.GetComponentsInChildren<Transform>();
+            return childs.Where(obj => obj.gameObject.name.Contains("audio") || obj.gameObject.name.Contains("SoundSrc")).ToArray();
         }
 
         /// <summary>
         /// Toggled uppon firewood being placed on the bed of flatbed.
         /// Switches the toggling method to UnityCar only, so the flatbed won't be unloaded completly.
         /// </summary>
-        internal void FlatbedSwitchToggleMethod()
+        void FlatbedSwitchToggleMethod()
         {
-            if (flatbedScriptActivated) return;
-            flatbedScriptActivated = true;
+            if (flatbedUnloadPreventionActivated) return;
+            flatbedUnloadPreventionActivated = true;
 
             Toggle = ToggleUnityCar;
         }
 
-        internal void RopeHookUp()
+        void RopeHookUp()
         {
             IsRopeHooked = true;
         }
 
-        internal void RopeUnhook()
+        void RopeUnhook()
         {
             IsRopeHooked = false;
         }