+++ MOP/src/MOP.cs
@@ -9,12 +9,14 @@ namespace MOP
         public override string ID => "MOP"; //Your mod ID (unique)
         public override string Name => "Modern Optimization Plugin"; //You mod name
         public override string Author => "Athlon"; //Your Username
-        public override string Version => "2.0.0"; //Version
+        public override string Version => "2.1.0"; //Version
 
         // Set this to true if you will be load custom assets from Assets folder.
         // This will create subfolder in Assets folder for your mod.
         public override bool UseAssetsFolder => false;
         public override bool SecondPass => true;
+
+        public static string ModConfigFolder;
 
         /// <summary>
         /// Called once, when mod is loading after game is fully loaded.
@@ -27,6 +29,8 @@ namespace MOP
                 ModUI.ShowMessage("MOP is not compatbile with KruFPS. Please remove KruFPS first!", "[MOP] Error");
                 return;
             }
+
+            ModConfigFolder = ModLoader.GetModConfigFolder(this);
         }
 
         /// <summary>
@@ -62,14 +66,14 @@ namespace MOP
         // ACTIVATING OBJECTS
         //
         public static Settings activeDistance = new Settings("activeDistance", "Active Distance", 1, MopSettings.UpdateAll);
-        // toggles
         public static Settings safeMode = new Settings("safeMode", "Safe Mode (requires restart)", false, MopSettings.UpdateAll);
+        public static Settings ignoreModVehicles = new Settings("ignoreModVehicles", "Ignore Mod Vehicles", false, MopSettings.UpdateAll);
 
         //
         // OTHERS
         //
-        public static Settings removeEmptyBeerBottles = new Settings("removeEmptyBeerBottles", "Remove Empty Beer Bottles", false);
-        public static Settings satsumaTogglePhysicsOnly = new Settings("satsumaTogglePhysicsOnly", "SATSUMA: Toggle Physics Only", false);
+        public static Settings removeEmptyBeerBottles = new Settings("removeEmptyBeerBottles", "Remove Empty Beer Bottles", false, MopSettings.UpdateAll);
+        public static Settings satsumaTogglePhysicsOnly = new Settings("satsumaTogglePhysicsOnly", "SATSUMA: Toggle Physics Only", false, MopSettings.UpdateAll);
         Settings temporarilyDisablePhysicsToggling = new Settings("temporarilyDisablePhysicsToggling", "Temporarily Disable Physics Toggling", MopSettings.DisablePhysicsToggling);
 
         // 
@@ -89,10 +93,11 @@ namespace MOP
         public static Settings occlusionDouble = new Settings("occlusionDouble", "Fancy", false, MopSettings.UpdateAll);
 
         //
-        // CHANGELOG
+        // ADVANCED
         //
         public static Settings toggleVehicles = new Settings("toggleVehicles", "Vehicles", true, MopSettings.UpdateAll);
         public static Settings toggleItems = new Settings("toggleItems", "Shop Items", true, MopSettings.UpdateAll);
+        public static Settings toggleVehiclePhysicsOnly = new Settings("toggleVehiclePhysicsOnly", "Toggle Vehicles Physics Only", false, MopSettings.UpdateAll);
 
         readonly Color32 headerColor = new Color32(29, 29, 29, 255);
 
@@ -115,6 +120,7 @@ namespace MOP
             Settings.AddCheckBox(this, safeMode);
             Settings.AddText(this, "Safe Mode will only allow to toggle objects that are known to not to cause any issues.\n" +
                 "Note: framerate boost will be dramatically decreased!");
+            Settings.AddCheckBox(this, ignoreModVehicles);
 
             // Others
             Settings.AddHeader(this, "Other", headerColor);
@@ -154,6 +160,7 @@ namespace MOP
             Settings.AddText(this, "If unchecked, the following objects will not get disabled.\n" +
                 "WARNING: Disabling Vehicles without disabling toggled items may cause items to fall through on the ground. " +
                 "DO NOT disable any of these, unless you REALLY need to!");
+            Settings.AddCheckBox(this, toggleVehiclePhysicsOnly);
 
             // Changelog
             Settings.AddHeader(this, "Changelog", headerColor);